+++ src/run/mod.rs
@@ -27,6 +27,7 @@ use comfy_table::{
     Cell,
     Table,
 };
+use log::debug;
 use std::{
     convert::TryInto,
     io::Write,
@@ -279,9 +280,12 @@ pub fn default_format(display: &TableDisplay, entries: Vec<Entry>) -> Result<(),
 }
 
 pub fn zsh_add_history(command: String, socket_path: PathBuf) -> Result<(), Error> {
-    let data = CommandStart::from_env(command)?;
-
-    client::new(socket_path).send(&Message::CommandStart(data))?;
+    if command.starts_with(' ') {
+        debug!("not recording a command starting with a space")
+    } else {
+        let data = CommandStart::from_env(command)?;
+        client::new(socket_path).send(&Message::CommandStart(data))?;
+    }
 
     Ok(())
 }