+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   return code that is not 0.
 * Add option `--find-status`. When specified will find all commands
   with the given return code.
+* Ignore commands starting with ' ' (space). This should make it
+  easier to not record sensitive commands.
 
 ## 1.0.0 [2021-06-01]
 * No big changes just updated the dependencies.